CTR ( Click Through Rate )
A company wants to know the CTR ( Click Through Rate ) in order to identify whether spending their money on digital advertising is worth or not. A higher CTR represents more interest in that specific campaign, whereas a lower CTR can show that the ad may not be as relevant. High CTRs are important because they show that more people are clicking through the website. Along with this high CTRs also help to get better ad position for less money on online platforms like Google, Bing etc.